misc/HtmlTestSuite/results/test-1347.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml"><head><script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
<script src="/preload/resources/preload_helper.js"/>
<script>
    var t = async_test('Makes sure that Link headers preload resources and block window.onload after resource discovery');
</script>
</head><body>
<style>
    #background {
        width: 200px;
        height: 200px;
        background-image: url(resources/square.png?background);
    }
</style>
<link href="resources/dummy.css?link-header-preload-delay-onload" rel="stylesheet"/>
<script src="resources/dummy.js?link-header-preload-delay-onload"/>
<div id="background"/>
<script>
    document.write('&lt;img src="resources/square.png?link-header-preload-delay-onload"&gt;');
    window.addEventListener("load", t.step_func(function() {
        verifyPreloadAndRTSupport();
        var entries = performance.getEntriesByType("resource");
        var found_background_first = false;
        for (var i = 0; i &lt; entries.length; ++i) {
            var entry = entries[i];
            if (entry.name.indexOf("square") != -1) {
                if (entry.name.indexOf("background") != -1)
                    found_background_first = true;
                break;
            }
        }
        assert_true(found_background_first);
        verifyLoadedAndNoDoubleDownload("resources/square.png?link-header-preload-delay-onload");
        verifyLoadedAndNoDoubleDownload("resources/square.png?background");
        verifyLoadedAndNoDoubleDownload("resources/dummy.js?link-header-preload-delay-onload");
        verifyLoadedAndNoDoubleDownload("resources/dummy.css?link-header-preload-delay-onload");
        t.done();
    }));
</script>
</body></html>
